requirements.txt
feedparser==6.0.11
firebase-admin==6.5.0
python-dateutil==2.9.0.post0
